# E-Cart_WebApp/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "E Cart Web API",
    "description": "A Simple E-Cart(E-Commerce App) Web API",
    "version": "v1"
  },
  "paths": {
    "/api/Products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Gets a list of all products.",
        "operationId": "GetProducts",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Creating a product.",
        "operationId": "PostProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    },
    "/api/Products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Gets a specific product by ID.",
        "operationId": "GetProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Updating a product.",
        "operationId": "PutProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Deleting a product.",
        "operationId": "DeleteProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Category": {
        "required": [
          "categoryName"
        ],
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "integer",
            "format": "int32"
          },
          "categoryName": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "picture": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Customer": {
        "required": [
          "companyName"
        ],
        "type": "object",
        "properties": {
          "customerId": {
            "maxLength": 5,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "companyName": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string"
          },
          "contactName": {
            "maxLength": 30,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "contactTitle": {
            "maxLength": 30,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "address": {
            "maxLength": 60,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "city": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "region": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "country": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "phone": {
            "maxLength": 24,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "fax": {
            "maxLength": 24,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            },
            "nullable": true
          },
          "customerTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerDemographic"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CustomerDemographic": {
        "type": "object",
        "properties": {
          "customerTypeId": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "customerDesc": {
            "type": "string",
            "nullable": true
          },
          "customers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Customer"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Employee": {
        "required": [
          "firstName",
          "lastName"
        ],
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "integer",
            "format": "int32"
          },
          "lastName": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string"
          },
          "firstName": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string"
          },
          "title": {
            "maxLength": 30,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "titleOfCourtesy": {
            "maxLength": 25,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "hireDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "address": {
            "maxLength": 60,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "city": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "region": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "country": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "homePhone": {
            "maxLength": 24,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "extension": {
            "maxLength": 4,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "photo": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "reportsTo": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "photoPath": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "reportsToNavigation": {
            "$ref": "#/components/schemas/Employee"
          },
          "inverseReportsToNavigation": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Employee"
            },
            "nullable": true
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            },
            "nullable": true
          },
          "territories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Territory"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Order": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer",
            "format": "int32"
          },
          "customerId": {
            "maxLength": 5,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "employeeId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "orderDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "requiredDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "shippedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "shipVia": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "freight": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "shipName": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "shipAddress": {
            "maxLength": 60,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "shipCity": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "shipRegion": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "shipPostalCode": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "shipCountry": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "employee": {
            "$ref": "#/components/schemas/Employee"
          },
          "shipViaNavigation": {
            "$ref": "#/components/schemas/Shipper"
          },
          "orderDetails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderDetail"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OrderDetail": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer",
            "format": "int32"
          },
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "unitPrice": {
            "type": "number",
            "format": "double"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "discount": {
            "type": "number",
            "format": "float"
          },
          "order": {
            "$ref": "#/components/schemas/Order"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          }
        },
        "additionalProperties": false
      },
      "Product": {
        "required": [
          "productName"
        ],
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "productName": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string"
          },
          "supplierId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "categoryId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "quantityPerUnit": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "unitPrice": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "unitsInStock": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "unitsOnOrder": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "reorderLevel": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "discontinued": {
            "type": "boolean"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "supplier": {
            "$ref": "#/components/schemas/Supplier"
          },
          "orderDetails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderDetail"
            },
            "nullable": true
          },
          "productsCarts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductsCart"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductsCart": {
        "type": "object",
        "properties": {
          "productCardId": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "productId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "cartQuantity": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "user": {
            "$ref": "#/components/schemas/UserMaster"
          }
        },
        "additionalProperties": false
      },
      "Region": {
        "required": [
          "regionDescription"
        ],
        "type": "object",
        "properties": {
          "regionId": {
            "type": "integer",
            "format": "int32"
          },
          "regionDescription": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "territories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Territory"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RoleMaster": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "integer",
            "format": "int32"
          },
          "roleDescription": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "userMasters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserMaster"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Shipper": {
        "required": [
          "companyName"
        ],
        "type": "object",
        "properties": {
          "shipperId": {
            "type": "integer",
            "format": "int32"
          },
          "companyName": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string"
          },
          "phone": {
            "maxLength": 24,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Supplier": {
        "required": [
          "companyName"
        ],
        "type": "object",
        "properties": {
          "supplierId": {
            "type": "integer",
            "format": "int32"
          },
          "companyName": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string"
          },
          "contactName": {
            "maxLength": 30,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "contactTitle": {
            "maxLength": 30,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "address": {
            "maxLength": 60,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "city": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "region": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "country": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "phone": {
            "maxLength": 24,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "fax": {
            "maxLength": 24,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "homePage": {
            "type": "string",
            "nullable": true
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Territory": {
        "required": [
          "territoryDescription"
        ],
        "type": "object",
        "properties": {
          "territoryId": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "territoryDescription": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "regionId": {
            "type": "integer",
            "format": "int32"
          },
          "region": {
            "$ref": "#/components/schemas/Region"
          },
          "employees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Employee"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserMaster": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "userName": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "password": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "nullable": true
          },
          "role": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "roleNavigation": {
            "$ref": "#/components/schemas/RoleMaster"
          },
          "productsCarts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductsCart"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}